Move pitch displacement helper into homography; prefer forward RANSAC.
Keeps SPEED independent of pass_pitch and matches the Copilot RANSAC note.
